Update questions.md
Hi Inanc Gumus,
Thank you for the helpful course into Golang, I'm trying to dive into the language starting with your material. By doing so, I found an answer to the question that might have an error, it has an additional backslash. I tried to compile it to confirm but it prints out an error.

Answer: "\"Hello\\""
Proposed answer: "\"Hello\""